+++ packages/openapi-ts/src/plugins/@hey-api/transformers/index.ts
@@ -1,2 +1,4 @@
+export { compiler } from '../../../compiler';
 export { defaultConfig, defineConfig } from './config';
+export { type expressionTransformer } from './expressions-transformers';
 export type { HeyApiTransformersPlugin } from './types';
